- Some warning fixes. - Added Planeshift.Client.Loading.UnloadLast option to psclient.cfg which   alters whether unloading is done before or after map loading. Set to true   by default, maps are loaded before old are unloaded (faster for machines   with enough ram to not swap, possibly slower for those who swap). Will   add GUI option in future.

+++ src/common/effects/pseffectobjlabel.cpp
@@ -378,8 +378,8 @@ bool psEffectObjLabel::SetText(int rows, ...)
     int mw,mh;
     fact->GetMaterialWrapper()->GetMaterial()->GetTexture()->GetOriginalDimensions(mw, mh);
 
-    facState->SetVertexCount( lettercount * 4 );
-    facState->SetTriangleCount( lettercount * 2 );
+    facState->SetVertexCount( (int)lettercount * 4 );
+    facState->SetTriangleCount( (int)lettercount * 2 );
 
     size_t elementCount = elemBuffer.GetSize();
     int cp = 0;
